move buildworkspace -> apply
